+++ R/parse_jdx.R
@@ -7,12 +7,17 @@
 #'
 #' @inherit lr_parse_generic return
 #'
+#' @details
+#' 'processed' column computed by lightr with the function
+#' [compute_processed()].
+#'
 #' @references McDonald RS, Wilks PA. JCAMP-DX: A Standard Form for Exchange of
 #' Infrared Spectra in Computer Readable Form. Applied Spectroscopy.
 #' 1988;42(1):151-62.
 #'
 #' @examples
-#' lr_parse_jdx(system.file("testdata", "OceanOptics.jdx", package = "lightr"))
+#' lr_parse_jdx(system.file("testdata", "OceanOptics_period.jdx",
+#'                          package = "lightr"))
 #'
 #' @export
 #'
@@ -20,9 +25,9 @@ lr_parse_jdx <- function(filename) {
   content <- readLines(filename)
   author <- grep("^##OWNER=", content, value = TRUE)
   author <- gsub("^##OWNER= ", "", author)[1]
-  savetime <- NA
-  specmodel <- NA
-  specID <- NA
+  savetime <- NA_character_
+  specmodel <- NA_character_
+  specID <- NA_character_
 
   # According to the standard, all blocks must start and end in this way:
   blockstarts <- grep("^##TITLE=", content)[-1]
@@ -74,8 +79,10 @@ lr_parse_jdx <- function(filename) {
     # Data is contained in lines that do NOT start with ##
     data <- grep("^##", content[blockstarts[index]:blockends[index]],
                  value = TRUE, invert = TRUE)
-    data <- strsplit(data, ",")
+    data <- strsplit(data, ", ")
     data <- do.call(rbind, data)
+    # Fix decimal for non-English files
+    data <- gsub(",", ".", data)
   }
 
   scope_data <- get_data(which(blocktype=="processed"))
@@ -84,8 +91,11 @@ lr_parse_jdx <- function(filename) {
 
   data <- cbind(scope_data[,1], dark_data[,2], white_data[,2], scope_data[,2])
   colnames(data) <- c("wl", "dark", "white", "scope")
-  data <- data.frame(apply(data, 2, as.numeric))
-  data$processed <- (data$scope - data$dark) / (data$white - data$dark) * 100
+
+  storage.mode(data) <- "numeric"
+  data <- as.data.frame(data)
+
+  data$processed <- compute_processed(data)
 
   return(list(data, metadata))