+++ tests/testthat/test_dispatch.R
@@ -8,8 +8,8 @@ test_that("Fallback", {
   )
 
   expect_equal(
-    lr_parse_jdx(test.file("OceanOptics.jdx")),
-    dispatch_parser(test.file("OceanOptics.jdx"))
+    lr_parse_jdx(test.file("OceanOptics_period.jdx")),
+    dispatch_parser(test.file("OceanOptics_period.jdx"))
   )
 
   expect_equal(
@@ -42,21 +42,33 @@ test_that("Fallback", {
     dispatch_parser(test.file("avantes_export2.trt"))
   )
 
+  expect_equal(
+    lr_parse_spc(test.file("compare/CRAIC/CRAIC.spc")),
+    dispatch_parser(test.file("compare/CRAIC/CRAIC.spc"))
+  )
+
+  expect_equal(
+    lr_parse_rfl8(test.file("compare", "Avantes", "feather.RFL8"), specnum = 1),
+    dispatch_parser(test.file("compare", "Avantes", "feather.RFL8"), specnum = 1)
+  )
+
 })
 
 test_that("Similar output for all parsers", {
 
-  files <- list.files(system.file("testdata", package = "lightr"),
+  files <- list.files(test.file(),
                       recursive = TRUE, include.dirs = TRUE)
-  files <- files[!tools::file_ext(files) %in% c("", "fail")]
+  files <- files[!tools::file_ext(files) %in% c("", "fail", "DRK", "REF")]
 
   lapply(files, function(file) {
-    res <- expect_silent(dispatch_parser(test.file(file), sep = ","))
+    res <- expect_silent(dispatch_parser(test.file(file), sep = ",", specnum = 1))
     expect_length(res, 2)
     expect_is(res[[1]], "data.frame")
     expect_true(all(apply(res[[1]], 2, is.numeric)))
     expect_named(res[[1]], c("wl", "dark", "white", "scope", "processed"))
     expect_length(res[[2]], 13)
+    expect_is(res[[2]], "character")
+    expect_true(is.na(res[[2]][2]) || grepl("^\\d{4}\\-[01]\\d-[0123]\\d$", res[[2]][2]))
   })
 
 })