+++ tests/testthat/test_multicores.R
@@ -2,10 +2,14 @@ context("multicore")
 
 test_that("Multicores/single core", {
 
-  expect_equivalent(lr_get_spec(test.file(), ext = "jdx", cores = 2),
-                    lr_get_spec(test.file(), ext = "jdx"))
+  expect_warning(lr_get_spec(test.file(), ext = "test", cores = 2),
+                 "deprecated")
 
-  expect_equivalent(lr_get_metadata(test.file(), ext = "jdx", cores = 2),
-                    lr_get_metadata(test.file(), ext = "jdx"))
+  expect_warning(lr_get_metadata(test.file(), ext = "test", cores = 2),
+                 "deprecated")
+
+  # Be careful here because lr_convert_tocsv() leaves breadcrumbs
+  expect_warning(lr_convert_tocsv(test.file(), ext = "test", cores = 2),
+                 "deprecated")
 
 })